Implementing JUnit Tests for Parse Class
